+++ src/main/java/bio/terra/service/cohortbuilder/InstancesService.java
@@ -11,12 +11,12 @@ import bio.terra.model.Literal;
 import bio.terra.model.LiteralValueUnion;
 import bio.terra.model.Query;
 import bio.terra.model.ValueDisplay;
-import bio.terra.service.cohortbuilder.tanagra.FromApiConversionService;
-import bio.terra.service.cohortbuilder.tanagra.instances.EntityInstanceCount;
-import bio.terra.service.cohortbuilder.tanagra.instances.filter.EntityFilter;
-import bio.terra.service.cohortbuilder.tanagra.service.QuerysService;
-import bio.terra.service.cohortbuilder.tanagra.service.UnderlaysService;
-import bio.terra.service.cohortbuilder.tanagra.utils.ToApiConversionUtils;
+import bio.terra.tanagra.service.FromApiConversionService;
+import bio.terra.tanagra.service.QuerysService;
+import bio.terra.tanagra.service.UnderlaysService;
+import bio.terra.tanagra.service.instances.EntityInstanceCount;
+import bio.terra.tanagra.service.instances.filter.EntityFilter;
+import bio.terra.tanagra.service.utils.ToApiConversionUtils;
 import bio.terra.tanagra.underlay.Attribute;
 import bio.terra.tanagra.underlay.Underlay;
 import java.util.ArrayList;
