--- src/main/java/bio/terra/service/cohortbuilder/tanagra/instances/filter/EntityFilter.java
@@ -1,10 +1,0 @@
-package bio.terra.service.cohortbuilder.tanagra.instances.filter;
-
-import bio.terra.tanagra.query.FilterVariable;
-import bio.terra.tanagra.query.TableVariable;
-import java.util.List;
-
-public abstract class EntityFilter {
-  public abstract FilterVariable getFilterVariable(
-      TableVariable entityTableVar, List<TableVariable> tableVars);
-}
